rename: 파일 분리 후, 파일 이름 변경
