+++ src/pages/MainPage.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 // import { useState } from 'react';
 
-import { FestivalListType } from 'types/api/catergory';
+import { FestivalListType } from 'types/api/festival';
 
 //components
 import MainSlider from '@components/main/MainSlider';
