+++ src/types/api/festival.ts
@@ -1,0 +1,17 @@
+/** 2023/07/04 - 축제 리스트 데이터 Type - by sineTlsl */
+export interface FestivalListType {
+  festivalId: number;
+  status: string;
+  title: string;
+  image: string;
+  address: string;
+  category: string;
+  eventstartdate: string;
+  eventenddate: string;
+  reviewRating: number;
+  reviewCount: number;
+  likeCount: number;
+  area: string;
+  mapx: number;
+  mapy: number;
+}
