+++ src/types/api/index.ts
@@ -1,4 +1,5 @@
 export * from './category';
+export * from './festival';
 
 /** 2023/07/04 - 페이지 데이터 Type - by sineTlsl */
 export interface PageInfoType {
